Increased the speed of the loader
Instead of connecting every time to the database all user data is saved to local save file on load, so it only have to connect to the database once and can get everything with one sql query. This also allows for offline access since a connection to the database is not needed everytime since all data is stored locally.